Move the logic about creating APIs to module 'server.application' from 'server.mock'.
