+++ pymock_api/server/mock.py
@@ -3,12 +3,9 @@
 This module provides objects for mocking APIs as a web application with different Python framework.
 """
 
-import json
-import os
-from typing import Any, List, Union
+from typing import Any
 
 from .._utils import load_config
-from ..exceptions import FileFormatNotSupport
 from ..model.api_config import APIConfig, MockAPIs
 from .application import BaseAppServer, FlaskServer
 
@@ -51,9 +48,7 @@ class MockHTTPServer:
     @property
     def web_app(self) -> Any:
         """:obj:`Any`: Property with only getter for the instance of web application, e.g., *Flask*, *FastAPI*, etc."""
-        if not self._web_application:
-            self._web_application = self._app_server.setup()
-        return self._web_application
+        return self._app_server.web_application
 
     def create_apis(self, mocked_apis: MockAPIs) -> None:
         """Initial and create all mocked APIs from the data objects which be generated by configuration.
@@ -66,86 +61,4 @@ class MockHTTPServer:
             None
 
         """
-        for api_name, api_config in mocked_apis.apis.items():
-            # pylint: disable=exec-used
-            exec(
-                f"""def {api_name}() -> Union[str, dict]:
-                return _HTTPResponse.generate(data='{api_config.http.response.value}')
-            """
-            )
-            # pylint: disable=exec-used
-            exec(
-                f"""self.web_app.route(
-                    "{mocked_apis.base.url}{api_config.url}", methods=["{api_config.http.request.method}"]
-                )({api_name}) """
-            )
-
-
-class _HTTPResponse:
-    """*Data processing of HTTP response for mocked HTTP application*
-
-    Handle the HTTP response value from the mocked APIs configuration.
-    """
-
-    valid_file_format: List[str] = ["json"]
-
-    @classmethod
-    def generate(cls, data: str) -> Union[str, dict]:
-        """Generate the HTTP response by the data. It would try to parse it as JSON format data in the beginning. If it
-        works, it returns the handled data which is JSON format. But if it gets fail, it would change to check whether
-        it is a file path or not. If it is, it would search and read the file to get the content value and parse it to
-        return. If it isn't, it returns the data directly.
-
-        Args:
-            data (str): The HTTP response value.
-
-        Returns:
-            A string type or dict type value.
-
-        """
-        try:
-            return json.loads(data)
-        except:  # pylint: disable=broad-except, bare-except
-            if cls._is_file(path=data):
-                return cls._read_file(path=data)
-        return data
-
-    @classmethod
-    def _is_file(cls, path: str) -> bool:
-        """Check whether the data is a file path or not.
-
-        Args:
-            path (str): A string type value.
-
-        Returns:
-            It returns ``True`` if it is a file path and the file exists, nor it returns ``False``.
-
-        """
-        path_sep_by_dot = path.split(".")
-        path_sep_by_dot_without_non = list(filter(lambda e: e, path_sep_by_dot))
-        if len(path_sep_by_dot_without_non) > 1:
-            support = path_sep_by_dot[-1] in cls.valid_file_format
-            if not support:
-                raise FileFormatNotSupport(cls.valid_file_format)
-            return support
-        else:
-            return False
-
-    @classmethod
-    def _read_file(cls, path: str) -> dict:
-        """Read file by the path to get its content and parse it as JSON format value.
-
-        Args:
-            path (str): The file path which records JSON format value.
-
-        Returns:
-            A dict type value which be parsed from JSON format value.
-
-        """
-        exist_file = os.path.exists(path)
-        if not exist_file:
-            raise FileNotFoundError(f"The target configuration file {path} doesn't exist.")
-
-        with open(path, "r", encoding="utf-8") as file_stream:
-            data = file_stream.read()
-        return json.loads(data)
+        self._app_server.create_api(mocked_apis)
